requirements.txt
annotated-types==0.7.0
anyio==4.3.0
browser-cookie3==0.19.1
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
decorator==4.4.2
distro==1.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.4.9
lz4==4.3.3
moviepy==1.0.3
numpy==1.26.4
openai==1.30.2
opencv-python==4.9.0.80
pillow==10.3.0
proglog==0.1.10
pycryptodomex==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
python-dotenv==1.0.1
requests==2.32.2
setuptools==70.0.0
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.12.0
urllib3==2.2.1
